+++ docs/README.md
@@ -1,4 +1,4 @@
-# Geegar User Guide
+# geegar.Geegar User Guide
 
 // Update the title above to match the actual product name
