+++ src/main/java/geegar/Geegar.java
@@ -1,40 +1,11 @@
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Scanner;
-import java.util.ArrayList;
+package geegar;
 
-//enum Command {
-//    LIST("list"),
-//    MARK("mark"),
-//    UNMARK("unmark"),
-//    TODO("todo"),
-//    DEADLINE("deadline"),
-//    EVENT("event"),
-//    DELETE("delete"),
-//    BYE("bye"),
-//    SCHEDULE("on");
-//
-//    private final String keyword;
-//
-//    Command (String keyword) {
-//        this.keyword = keyword;
-//    }
-//
-//    public String getKeyword() {
-//        return this.keyword;
-//    }
-//
-//    public static Command convertInput(String input) {
-//        String command = input.toLowerCase().split(" ")[0];
-//        for (Command cmd : Command.values()) {
-//            if (cmd.getKeyword().equals(command)) {
-//                return cmd;
-//            }
-//        }
-//        return null;
-//    }
-//}
+import geegar.command.Command;
+import geegar.exception.GeegarException;
+import geegar.parser.Parser;
+import geegar.task.TaskList;
+import geegar.ui.Ui;
+import geegar.storage.Storage;
 
 public class Geegar {
     private Storage storage;
@@ -78,10 +49,10 @@ public class Geegar {
 }
 //    private static final int UNDERSCORE_LENGTH = 60;
 //    private static final String OGRE_EMOJI = "\uD83E\uDDCC";
-//    private static ArrayList<Task> taskList = new ArrayList<>();
+//    private static ArrayList<geegar.task.Task> taskList = new ArrayList<>();
 //
 //    public static void main(String[] args) {
-//        taskList = TaskReader.loadTasks();
+//        taskList = geegar.storage.TaskReader.loadTasks();
 //
 //        printIntroduction();
 //        Scanner sc = new Scanner(System.in);
@@ -90,16 +61,16 @@ public class Geegar {
 //        while (true) {
 //            String input = sc.nextLine();
 //
-//            if (input.equalsIgnoreCase(Command.BYE.getKeyword())) {
+//            if (input.equalsIgnoreCase(geegar.command.Command.BYE.getKeyword())) {
 //                break;
 //            }
 //
 //            try {
-//                Command command = Command.convertInput(input);
+//                geegar.command.Command command = geegar.command.Command.convertInput(input);
 //
 //                if (command == null) {
 //                    if (!input.isEmpty()) {
-//                        throw new UnknownCommandException(input);
+//                        throw new geegar.exception.UnknownCommandException(input);
 //                    }
 //                    continue;
 //                }
@@ -130,10 +101,10 @@ public class Geegar {
 //                        handleOnCommand(input);
 //                        break;
 //                    default:
-//                        throw new UnknownCommandException(input);
+//                        throw new geegar.exception.UnknownCommandException(input);
 //                }
 //
-//            } catch (GeegarException e) {
+//            } catch (geegar.exception.GeegarException e) {
 //                printError(e.getMessage());
 //            }
 ////            handleAddTask(input);
@@ -146,7 +117,7 @@ public class Geegar {
 //
 //    private static void printIntroduction() {
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
-//        System.out.println("Hello! I'm Geegar " + OGRE_EMOJI);
+//        System.out.println("Hello! I'm geegar.Geegar " + OGRE_EMOJI);
 //        System.out.println("What can I do for you today?");
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //    }
@@ -174,7 +145,7 @@ public class Geegar {
 //    }
 //
 ////    private static void handleAddTask(String input) {
-////        taskList[index] = new Task(input);
+////        taskList[index] = new geegar.task.Task(input);
 ////        index++;
 ////
 ////        System.out.println("_".repeat(UNDERSCORE_LENGTH));
@@ -183,28 +154,28 @@ public class Geegar {
 ////    }
 //
 //
-//    private static void handleMarkCommand(String input) throws InvalidTaskNumberException {
+//    private static void handleMarkCommand(String input) throws geegar.exception.InvalidTaskNumberException {
 //        String[] parts = input.split(" ");
 //        int taskNumber = Integer.parseInt(parts[1]);
 //        if (taskNumber < 1 || taskNumber > taskList.size()) {
-//            throw new InvalidTaskNumberException(parts[1]);
+//            throw new geegar.exception.InvalidTaskNumberException(parts[1]);
 //        }
 //        taskList.get(taskNumber - 1).markAsDone();
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Nice! I've marked this task as done:");
 //        System.out.println(taskList.get(taskNumber - 1));
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //    }
 //
-//    private static void handleUnmarkCommand(String input) throws InvalidTaskNumberException {
+//    private static void handleUnmarkCommand(String input) throws geegar.exception.InvalidTaskNumberException {
 //        String[] parts = input.split(" ");
 //        int taskNumber = Integer.parseInt(parts[1]);
 //        if (taskNumber < 1 || taskNumber > taskList.size()) {
-//            throw new InvalidTaskNumberException(parts[1]);
+//            throw new geegar.exception.InvalidTaskNumberException(parts[1]);
 //        }
 //        taskList.get(taskNumber - 1).markNotDone();
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Alright! I've marked this task as not done yet:");
 //        System.out.println(taskList.get(taskNumber - 1));
@@ -212,16 +183,16 @@ public class Geegar {
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //    }
 //
-//    private static void handleTodoCommand(String input) throws GeegarException {
+//    private static void handleTodoCommand(String input) throws geegar.exception.GeegarException {
 //        if (input.length() <= 5) {
-//            throw new EmptyDescriptionException("todo");
+//            throw new geegar.exception.EmptyDescriptionException("todo");
 //        }
 //        String description = input.substring(5); // remove the todo keyword from input
 //
-//        Todo newTask = new Todo(description);
+//        geegar.task.Todo newTask = new geegar.task.Todo(description);
 //        taskList.add(newTask);
 //
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Got it. I've added this task:");
@@ -231,33 +202,33 @@ public class Geegar {
 //
 //    }
 //
-//    private static void handleDeadlineCommand(String input) throws GeegarException {
+//    private static void handleDeadlineCommand(String input) throws geegar.exception.GeegarException {
 //        if (input.length() <= 9) {
-//            throw new EmptyDescriptionException("deadline");
+//            throw new geegar.exception.EmptyDescriptionException("deadline");
 //        }
 //        String content = input.substring(9); // remove the deadline keyword from input
 //        if (!content.contains( " /by ")) {
-//            throw new InvalidFormatDeadlineException();
+//            throw new geegar.exception.InvalidFormatDeadlineException();
 //        }
 //        String[] parts = content.split(" /by ");
 //
 //        String description = parts[0].trim();
 //        if (description.isEmpty()) {
-//            throw new EmptyDescriptionException("Deadline");
+//            throw new geegar.exception.EmptyDescriptionException("geegar.task.Deadline");
 //        }
 //        String byInput = parts[1].trim();
 //
 //        if (byInput.isEmpty()) {
-//            throw new InvalidFormatDeadlineException();
+//            throw new geegar.exception.InvalidFormatDeadlineException();
 //        }
 //
 //        // Accept format "dd/MM/yyyy HHmm" like 27/08/2025 0600
 //        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
 //        LocalDateTime by = LocalDateTime.parse(byInput, formatter);
 //
-//        Deadline newTask = new Deadline(description, by);
+//        geegar.task.Deadline newTask = new geegar.task.Deadline(description, by);
 //        taskList.add(newTask);
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Got it. I've added this task:");
@@ -267,31 +238,31 @@ public class Geegar {
 //
 //    }
 //
-//    private static void handleEventCommand(String input) throws GeegarException {
-//        if (input.length() <= 6) { throw new EmptyDescriptionException("event"); }
+//    private static void handleEventCommand(String input) throws geegar.exception.GeegarException {
+//        if (input.length() <= 6) { throw new geegar.exception.EmptyDescriptionException("event"); }
 //        String content = input.substring(6); // remove the event keyword from input
 //
 //        if (!content.contains( " /from ")) {
-//            throw new InvalidFormatEventException();
+//            throw new geegar.exception.InvalidFormatEventException();
 //        }
 //
 //        String[] splitByFrom = content.split(" /from ");
 //        if (splitByFrom.length != 2) {
-//            throw new InvalidFormatEventException();
+//            throw new geegar.exception.InvalidFormatEventException();
 //        }
 //        if (!splitByFrom[1].contains(" /to ")) {
-//            throw new InvalidFormatEventException();
+//            throw new geegar.exception.InvalidFormatEventException();
 //        }
 //        String[] splitByTo = splitByFrom[1].split(" /to ");
 //
 //        String description = splitByFrom[0].trim();
 //        if (description.isEmpty()) {
-//            throw new EmptyDescriptionException("Event");
+//            throw new geegar.exception.EmptyDescriptionException("geegar.task.Event");
 //        }
 //        String fromInput = splitByTo[0].trim();
 //        String toInput = splitByTo[1].trim();
 //        if (fromInput.isEmpty() || toInput.isEmpty()) {
-//            throw new InvalidFormatEventException();
+//            throw new geegar.exception.InvalidFormatEventException();
 //        }
 //
 //        // Accept format "dd/MM/yyyy HHmm" like 27/08/2025 0600
@@ -299,9 +270,9 @@ public class Geegar {
 //        LocalDateTime from = LocalDateTime.parse(fromInput, formatter);
 //        LocalDateTime to = LocalDateTime.parse(toInput, formatter);
 //
-//        Event newTask = new Event(description, from, to);
+//        geegar.task.Event newTask = new geegar.task.Event(description, from, to);
 //        taskList.add(newTask);
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Got it. I've added this task:");
@@ -317,19 +288,19 @@ public class Geegar {
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //    }
 //
-//    private static void handleDeleteCommand(String input) throws GeegarException {
+//    private static void handleDeleteCommand(String input) throws geegar.exception.GeegarException {
 //        String[] parts = input.split(" ");
 //        int taskNumber = Integer.parseInt(parts[1]);
 //
 //        if (taskNumber < 1 || taskNumber > taskList.size()) {
-//            throw new InvalidTaskNumberException(parts[1]);
+//            throw new geegar.exception.InvalidTaskNumberException(parts[1]);
 //        }
 //
-//        Task deletedTask = taskList.get(taskNumber - 1);
+//        geegar.task.Task deletedTask = taskList.get(taskNumber - 1);
 //        taskList.remove(taskNumber - 1);
 //
 //        // Writes into the txt file the entire file again with deletion
-//        TaskWriter.saveTasks(taskList);
+//        geegar.storage.TaskWriter.saveTasks(taskList);
 //
 //        System.out.println("_".repeat(UNDERSCORE_LENGTH));
 //        System.out.println(OGRE_EMOJI + ": Got it. I've deleted this task:");
@@ -339,7 +310,7 @@ public class Geegar {
 //
 //    }
 //
-//    private static void handleOnCommand(String input) throws GeegarException {
+//    private static void handleOnCommand(String input) throws geegar.exception.GeegarException {
 //        String timeInput = input.substring(4);
 //        DateTimeFormatter inputFmt = DateTimeFormatter.ofPattern("d/M/yyyy");
 //        LocalDate date = LocalDate.parse(timeInput.trim(), inputFmt);
