+++ src/main/java/geegar/exception/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package geegar.exception;
+
+public class EmptyDescriptionException extends GeegarException {
+    public EmptyDescriptionException(String taskType) {
+        super("Hey! The description of a " + taskType + " cannot be empty.");
+    }
+}
