+++ src/main/java/geegar/exception/GeegarException.java
@@ -1,0 +1,7 @@
+package geegar.exception;
+
+public class GeegarException extends Exception {
+    public GeegarException(String message) {
+        super(message);
+    }
+}
