+++ src/main/java/geegar/exception/InvalidFormatDeadlineException.java
@@ -1,0 +1,9 @@
+package geegar.exception;
+
+public class InvalidFormatDeadlineException extends GeegarException {
+    public InvalidFormatDeadlineException() {
+        super("Invalid deadline format! Use: deadline <description> /by <time>");
+    }
+}
+
+
