+++ src/main/java/geegar/exception/InvalidFormatEventException.java
@@ -1,0 +1,8 @@
+package geegar.exception;
+
+public class InvalidFormatEventException extends GeegarException {
+    public InvalidFormatEventException() {
+        super("Invalid event format! Use: event <description> /from <start> /to <end>");
+
+    }
+}
