+++ src/main/java/geegar/exception/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package geegar.exception;
+
+public class InvalidTaskNumberException extends GeegarException {
+    public InvalidTaskNumberException(String taskNumber) {
+        super("geegar.task.Task Number does not exist!");
+    }
+}
