+++ src/main/java/geegar/exception/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package geegar.exception;
+
+public class UnknownCommandException extends GeegarException {
+    public UnknownCommandException(String command) {
+        super("I don't know what " + command + " means, try 'todo', 'deadline' or 'event' to add a task!");
+    }
+}
