+++ src/main/java/geegar/task/Deadline.java
@@ -1,3 +1,5 @@
+package geegar.task;
+
 import java.time.LocalDateTime;
 
 public class Deadline extends Task {
