+++ src/main/java/geegar/task/Event.java
@@ -1,5 +1,6 @@
+package geegar.task;
+
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 public class Event extends Task {
     protected LocalDateTime from;
