+++ src/main/java/geegar/task/Task.java
@@ -1,3 +1,5 @@
+package geegar.task;
+
 import java.time.format.DateTimeFormatter;
 
 public abstract class Task {
