+++ src/main/java/geegar/task/Todo.java
@@ -1,3 +1,5 @@
+package geegar.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
