+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Geegar 🧌
+Hello! I'm geegar.Geegar 🧌
 What can I do for you today?
 ____________________________________________________________
 ____________________________________________________________
